requirements.txt
python-dotenv
mcp
fastapi
autogen-agentchat
autogen-ext[openai]
autogen-ext[mcp]
mcp-server-fetch